fix: AB Trails Android flicker at rendering layer
