+++ src/shared/ui/SketchCanvas/SketchCanvas.tsx
@@ -2,18 +2,10 @@ import { forwardRef, useImperativeHandle } from 'react';
 import { StyleSheet, View } from 'react-native';
 
 import { Canvas, Path, Skia, SkPath } from '@shopify/react-native-skia';
-import { GestureDetector } from 'react-native-gesture-handler';
-import { runOnJS, useSharedValue } from 'react-native-reanimated';
+import { useSharedValue } from 'react-native-reanimated';
 
-import {
-  createLine,
-  createPathFromPoints,
-  getCurrentShape,
-  Point,
-  progressLine,
-  Shape,
-} from './LineSketcher';
-import { useDrawingGesture } from './useDrawingGesture';
+import { DrawingLayer } from './DrawingLayer';
+import { createPathFromPoints, Point } from './LineSketcher';
 
 export type SketchCanvasRef = {
   clear: () => void;
@@ -42,9 +34,6 @@ export const SketchCanvas = forwardRef<SketchCanvasRef, Props>((props, ref) => {
       .reduce((path, curr) => path.addPath(curr), Skia.Path.Make()),
   );
 
-  const points = useSharedValue<Point[]>([]);
-
-  const lastPointTime = useSharedValue<number | null>(null);
   const width = useSharedValue(0);
 
   useImperativeHandle(ref, () => {
@@ -55,94 +44,35 @@ export const SketchCanvas = forwardRef<SketchCanvasRef, Props>((props, ref) => {
     };
   });
 
-  const onTouchStart = (point: Point, time: number) => {
-    'worklet';
-    fullPath.value.addPath(createLine(points, point));
-    runOnJS(onStrokeStart)(point.x, point.y, time);
-  };
-
-  const onTouchProgress = (
-    point: Point,
-    straightLine: boolean,
-    time: number,
-  ) => {
-    'worklet';
-    const lastDrawnPoint = points.value[points.value.length - 1];
-
-    if (lastDrawnPoint) {
-      const dx = point.x - lastDrawnPoint.x;
-      const dy = point.y - lastDrawnPoint.y;
-
-      const isSameTime = lastPointTime.value === time;
-      const isSamePoint = dx === 0 && dy === 0;
-
-      if (isSamePoint || isSameTime) {
-        return;
-      }
-    }
-
-    lastPointTime.value = time;
-
-    fullPath.modify(value => {
-      'worklet';
-      progressLine(points, value, point, straightLine);
-
-      return value;
-    });
-
-    runOnJS(onStrokeChanged)(point.x, point.y, time);
-  };
-
-  const createDot = (point: Point, time: number) => {
-    'worklet';
-    runOnJS(onStrokeChanged)(point.x, point.y, time);
-
-    fullPath.modify(value => {
-      'worklet';
-      progressLine(points, value, point);
-
-      return value;
-    });
-  };
-
-  const onTouchEnd = (point: Point, time: number) => {
-    'worklet';
-    if (getCurrentShape(points) === Shape.Dot) {
-      const firstPoint = points.value[points.value.length - 1];
-
-      createDot(
-        {
-          x: firstPoint.x + 1.5,
-          y: firstPoint.y + 1.5,
-        },
-        Date.now(),
-      );
-    }
-
-    runOnJS(onStrokeEnd)(point.x, point.y, time);
-  };
-
-  const drawingGesture = useDrawingGesture(
-    { areaSize: width, enabled },
-    { onTouchStart, onTouchProgress, onTouchEnd },
-  );
-
   return (
-    <GestureDetector gesture={drawingGesture}>
-      <View
-        style={styles.canvas}
-        onLayout={e => (width.value = e.nativeEvent.layout.width)}
-      >
-        <Canvas style={StyleSheet.absoluteFill}>
-          <Path
-            path={fullPath}
-            strokeWidth={1.5}
-            color="black"
-            style="stroke"
-          />
-        </Canvas>
-      </View>
-    </GestureDetector>
+    <View
+      style={styles.canvas}
+      onLayout={e => (width.value = e.nativeEvent.layout.width)}
+    >
+      {/*
+        Always mounted: keeping the Skia Canvas alive avoids the GLSurface
+        re-initialisation that flashes (flickers) on Android when it is
+        unmounted/remounted (regression introduced by the Skia 2.4.x upgrade).
+      */}
+      <Canvas style={StyleSheet.absoluteFill}>
+        <Path path={fullPath} strokeWidth={1.5} color="black" style="stroke" />
+      </Canvas>
+
+      {/*
+        Conditionally mounted: unmounting the gesture layer on every hide/show
+        cycle gives the next stroke a fully reset RNGH handler + gesture state,
+        which is what prevents free-drawing, multitouch and stale start points.
+      */}
+      {enabled && (
+        <DrawingLayer
+          fullPath={fullPath}
+          width={width}
+          onStrokeStart={onStrokeStart}
+          onStrokeChanged={onStrokeChanged}
+          onStrokeEnd={onStrokeEnd}
+        />
+      )}
+    </View>
   );
 });
